feat: refactors TaskService
