+++ app/components/ClientTaskList/TaskItem/DeleteConfirmationModal.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import Modal from '../../Modal/Modal';
 import { useTaskStore } from '@/lib/store/task';
-import type { TaskWithTags } from '@/lib/services/task';
+import { TaskWithTags } from '@/lib/db';
 import TaskItemSkeleton from '../../TaskItemSkeleton';
 import { primaryButtonStyles, secondaryButtonStyles } from '@/lib/style';
 type DeleteConfirmationModalProps = {
