+++ app/components/ClientTaskList/TaskItem/TaskItem.tsx
@@ -11,7 +11,7 @@ import DeleteConfirmationModal from './DeleteConfirmationModal';
 import DueDateModal from './DueDateModal';
 import MeatballMenu from '../../MeatballMenu';
 import ToggleCompletionButton from './ToggleCompletionButton';
-import type { TaskWithTags } from '@/lib/services/task';
+import { TaskWithTags } from '@/lib/db';
 
 /**
  * A component to render a single task.
