+++ lib/services/index.ts
@@ -1,4 +1,4 @@
-import { TaskService } from './task';
+import { TaskService } from './Task/task';
 import { NoteService } from './note';
 import { TagsService } from './tags';
